requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
alembic>=1.7.1,<2.0.0
email-validator==1.1.3
qrcode==7.4.2
Pillow==10.1.0
pymysql==1.0.2
cryptography==3.4.7
requests==2.31.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
jieba==0.42.1 